requirements.txt
openai
psycopg2-binary
jsonschema
pytest
vcrpy
